bfa updates
Bfa map info is added, but doesn't draw properly.. Nx.Map:MoveZoneTiles
was written expecting a 4x3 grid which all old maps used, the new maps
are 15x10 grids ... if someone wants to tackle that before I can, it
would be great :P otherwise I'll keep working on it when I can.

Mouseover of bfa zones shows the right zone (or was when I was assuming
4x3), but there is no overlays to draw areas as I haven't found a way to
obtain that info yet.